vetapp-gcp/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ── Google Cloud ──────────────────────────────────────────────────────────────
google-cloud-storage==2.17.0

# ── HTTP clients ──────────────────────────────────────────────────────────────
httpx==0.27.0
requests==2.32.3

# ── Image processing ──────────────────────────────────────────────────────────
pillow==10.3.0
opencv-python-headless==4.10.0.84
numpy==1.26.4

# ── ML: PyTorch (CPU-only — keeps image under 2 GB) ──────────────────────────
# CPU wheels are fetched from PyTorch's own index (see pip.conf / Dockerfile).
torch==2.3.0
torchvision==0.18.0

# ── ML: TensorFlow (CPU-only) ─────────────────────────────────────────────────
tensorflow-cpu==2.16.1
